% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_annotation}
\alias{add_annotation}
\title{Add annotations to ggplot heatmap}
\usage{
add_annotation(
  ggp,
  annot_dim = c("rows", "cols"),
  annot_df,
  annot_pos,
  annot_size,
  annot_border_lwd = 0.5,
  annot_border_col = "grey"
)
}
\arguments{
\item{ggp}{\code{ggplot} object to add annotations to}

\item{annot_dim}{Dimension to add annotations to, either "rows" or "columns"}

\item{annot_df}{Data frame containing the annotations.
The first column must contain the labels used in the heatmap rows or columns.
Each of the other columns should contain corresponding annotations. The column names are used as labels in the legend.}

\item{annot_pos}{Positions of annotations. x positions if row annotations, y positions if column annotations.
Positions are given by the middle coordinate of the annotation cells.}

\item{annot_size}{Size of annotations, width of row annotations, height if column annotations.}

\item{annot_border_lwd}{Linewidth of border lines of annotation cells.}

\item{annot_border_col}{Colour of border lines of annotation cells.}
}
\description{
Add annotations to ggplot heatmap
}
