% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{add_dendrogram}
\alias{add_dendrogram}
\title{Add dendrogram to ggplot object.}
\usage{
add_dendrogram(ggp, dend_seg, dend_col = "black", dend_lwd = 0.3, dend_lty = 1)
}
\arguments{
\item{ggp}{ggplot object to add dendrogram to.}

\item{dend_seg}{Dendrogram segment data obtained from the \code{prepare_dendrogram} function.}

\item{dend_col}{String specifying colour of dendrogram (used if the colours have not been changed using other options).}

\item{dend_lwd}{Line width of dendrogram segments (used if not changed using other options).}

\item{dend_lty}{Line type of dendrogram (used if not changed using other options).}
}
\value{
A ggplot object with a dendrogram added.
}
\description{
Add dendrogram to ggplot object.
}
\keyword{internal}
