% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrhm.R
\name{add_pvalue_labels}
\alias{add_pvalue_labels}
\title{Helper function to add p-value labels for ggcorrhm}
\usage{
add_pvalue_labels(
  cor_mat_dat = NULL,
  cor_plt_dat,
  cor_plt_plt,
  mode,
  skip_diag = F,
  cell_labels,
  cell_label_p,
  cell_label_col,
  cell_label_size,
  cell_label_digits,
  p_thresholds,
  border_col,
  border_lwd,
  border_lty,
  show_legend,
  col_scale
)
}
\arguments{
\item{cor_mat_dat}{Correlation matrix data with p-values (leave as NULL if not computed).}

\item{cor_plt_dat}{Plot data from gghm.}

\item{cor_plt_plt}{Plot from gghm.}

\item{mode}{The plotting mode.}

\item{skip_diag}{If the diagonal should be skipped (to not draw on names).}

\item{cell_labels}{Logical, if labels should be drawn.}

\item{cell_label_col}{Cell label colours.}

\item{cell_label_size}{Cell label sizes.}

\item{cell_label_digits}{Cell label digits.}

\item{p_thresholds}{P-value thresholds.}

\item{border_col}{Cell border colours.}

\item{border_lwd}{Cell border line widths.}

\item{border_lty}{Cell border line types.}

\item{show_legend}{Vector indicating which legends should be shown.}
}
\value{
Plot with labels added.
}
\description{
Helper function to add p-value labels for ggcorrhm
}
\keyword{internal}
