% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{add_split_diag}
\alias{add_split_diag}
\title{Draw triangles in the diagonal to split the it between the two triangles of a square matrix.}
\usage{
add_split_diag(
  plt,
  x_long,
  lt,
  border_col = "grey",
  border_lwd = 0.1,
  border_lty = 1
)
}
\arguments{
\item{plt}{Plot object to add to.}

\item{x_long}{Plotting data for the diagonal.}

\item{lt}{Layout of the heatmap.}

\item{border_col}{Cell border colour.}

\item{border_lwd}{Cell border linewidth.}

\item{border_lty}{Cell border linetype.}
}
\value{
ggplot object with added triangles in the diagonal.
}
\description{
Draw triangles in the diagonal to split the it between the two triangles of a square matrix.
}
\keyword{internal}
