% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrhm.R
\name{add_text_geom}
\alias{add_text_geom}
\title{Helper function for adding the text geom depending on mode/cell_labels}
\usage{
add_text_geom(
  dat,
  plt,
  type = c("text_mode", "cell_label"),
  show_legend,
  skip_diag,
  cell_label_col,
  cell_label_size,
  border_col,
  border_lwd,
  border_lty,
  col_scale
)
}
\arguments{
\item{dat}{Plotting data.}

\item{plt}{Plot to add to.}

\item{type}{String saying if it is 'text_mode' or 'cell_label' that is being drawn (changes scaling of colours).}

\item{show_legend}{Legends to be shown.}

\item{skip_diag}{If the diagonal should be skipped.}

\item{cell_label_col}{Cell label colours.}

\item{cell_label_size}{Cell label sizes.}

\item{border_col}{Cell border colours.}

\item{border_lwd}{Cell border line widths.}

\item{border_lty}{Cell border line types.}
}
\value{
The plot with the text geom added.
}
\description{
Helper function for adding the text geom depending on mode/cell_labels
}
\keyword{internal}
