% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dendextend.R
\name{apply_dendextend}
\alias{apply_dendextend}
\title{Successively apply dendextend functions to a dendrogram.}
\usage{
apply_dendextend(dendro, dend_list)
}
\arguments{
\item{dendro}{Dendrogram object obtained from \code{stats::as.dendrogram}.}

\item{dend_list}{List specifying dendextend functions to apply. For usage see the details of \code{gg_corr_heatmap}.}
}
\description{
Successively apply dendextend functions to a dendrogram.
}
\keyword{internal}
