% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{check_annot_df}
\alias{check_annot_df}
\title{Check names of annotation}
\usage{
check_annot_df(annot_df, names_in, context)
}
\arguments{
\item{annot_df}{Annotation data frame (annot_rows_df, annot_cols_df).}

\item{names_in}{Names that exist in the data (rownames, colnames).}
}
\value{
Annotation data frame (rownames moved to column named '.names' if necessary).
}
\description{
Check names of annotation
}
\keyword{internal}
