% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_annot_names_deprecated}
\alias{check_annot_names_deprecated}
\title{Check annotation name parameters for deprecated usage.}
\usage{
check_annot_names_deprecated(
  new_params = NULL,
  old_params = NULL,
  context = c("rows", "cols")
)
}
\arguments{
\item{new_params}{New argument input.}

\item{old_params}{Deprecated argument input.}

\item{context}{Context (rows or cols).}
}
\value{
A string stating which strategy to use for drawing names.
}
\description{
Check annotation name parameters for deprecated usage.
}
\keyword{internal}
