% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_cell_labels}
\alias{check_cell_labels}
\title{Check cell labels if they are ok.}
\usage{
check_cell_labels(cell_labels, x_long)
}
\arguments{
\item{cell_labels}{The \code{cell_labels} input to \code{gghm()}.}

\item{x_long}{Long format input data.}
}
\value{
Long format data frame containing cell labels.
}
\description{
Check cell labels if they are ok.
}
\keyword{internal}
