% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dendrogram_pos.R
\name{check_dendrogram_pos}
\alias{check_dendrogram_pos}
\title{Check that dendrograms are positioned correctly}
\usage{
check_dendrogram_pos(dat, dend_dim = c("row", "col"), dend)
}
\arguments{
\item{dat}{Long format data for plotting.}

\item{dend_dim}{Dimension to which the dendrogram is added.}

\item{dend}{The dendrogram segments data frame.}
}
\value{
\code{dend} is returned as is if the positions are correct. Otherwise there is an error.
}
\description{
Check that dendrograms are positioned correctly
}
\keyword{internal}
