% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_layout}
\alias{check_layout}
\title{Check that layout and mode are correct}
\usage{
check_layout(layout, mode)
}
\arguments{
\item{layout}{Plot layout.}

\item{mode}{Plot mode.}
}
\value{
Error if incorrect layout or mode, otherwise nothing.
}
\description{
Check that layout and mode are correct
}
\keyword{internal}
