% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_logical}
\alias{check_logical}
\title{Check (supposed) logical values.}
\usage{
check_logical(..., list_allowed = FALSE, call = NULL)
}
\arguments{
\item{...}{Should be a single named argument, where the name is the variable name displayed in the error message.
The value is the (supposed) logical.}

\item{list_allowed}{Logical indicating if the argument is allowed to be a list. If TRUE each element will be checked.}

\item{call}{Call to use for the call in the error message (used in rlang::abort).
Default is rlang::caller_env() resulting in the function that called check_logical().}
}
\value{
Error if not logical or longer than 1, otherwise nothing.
}
\description{
Check (supposed) logical values.
}
\keyword{internal}
