% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster data using hierarchical clustering or use provided clustering.}
\usage{
cluster_data(
  cluster_input,
  mat,
  cluster_distance,
  cluster_method,
  dend_options = NULL
)
}
\arguments{
\item{cluster_input}{Either a logical indicating if data should be clustered, or a \code{hclust} or \code{dendrogram} object.}

\item{mat}{Matrix to cluster.}

\item{cluster_distance}{Distance metric for clustering.}

\item{cluster_method}{Clustering method for \code{hclust}.}

\item{dend_options}{List or functional sequence specifying \code{dendextend} functions to use.}
}
\value{
List containing the dendrogram and clustering objects.
}
\description{
Cluster data using hierarchical clustering or use provided clustering.
}
\keyword{internal}
