% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_dimension.R
\name{cluster_dimension}
\alias{cluster_dimension}
\title{Cluster data using hierarchical clustering or use provided clustering.}
\usage{
cluster_dimension(
  cluster_data,
  mat,
  cluster_distance,
  cluster_method,
  dend_options = NULL
)
}
\arguments{
\item{cluster_data}{Either a logical indicating if data should be clustered, or a \code{hclust} object.}

\item{mat}{Matrix to cluster.}

\item{cluster_distance}{Distance metric for clustering.}

\item{cluster_method}{Clustering method for \code{hclust}.}

\item{dend_options}{List specifying \code{dendextend} functions to use.}
}
\value{
List containing the dendrogram and clustering objects (or NULL if no clustering).
}
\description{
Cluster data using hierarchical clustering or use provided clustering.
}
\keyword{internal}
