% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghm_tidy.R
\name{cor_long}
\alias{cor_long}
\title{Make a correlation matrix from long format data.}
\usage{
cor_long(
  x,
  rows1,
  cols1,
  values1,
  y = NULL,
  rows2 = NULL,
  cols2 = NULL,
  values2 = NULL,
  out_format = c("wide", "long"),
  method = "pearson",
  use = "everything"
)
}
\arguments{
\item{x}{A long format data frame containing the data to correlate.}

\item{rows1, cols1}{The columns in \code{x} containing the values that should be in the rows and columns of the correlation matrix.}

\item{values1}{Name of the column in \code{x} containing the values of the correlation matrix.}

\item{y}{Optional second data frame for correlating with the data frame from \code{x}.}

\item{rows2, cols2}{Optional names of columns with values for the rows and columns of a second matrix. If \code{y} is a data frame, \code{rows2} is taken from \code{y}.
Otherwise it is taken from \code{x}.}

\item{values2}{Optional column for the values of a second matrix.}

\item{out_format}{Format of output correlation matrix ("long" or "wide").}

\item{method}{Correlation method given to \code{stats::cor()}.}

\item{use}{Missing value strategy of \code{stats::cor()}.}
}
\value{
A correlation matrix (if wide format) or a long format data frame with the columns
'row', 'col', and 'value' (containing correlations).
}
\description{
Make a correlation matrix from long format data.
}
\details{
If there is only one input data frame (\code{x}) and \code{rows2}, \code{cols2} and \code{values2} are NULL (the default),
a wide matrix is constructed from \code{x} and passed to \code{stats::cor()}, resulting in a correlation matrix
with the column-column correlations.

If \code{y} is a data frame and \code{rows2}, \code{cols2} and \code{values2} are specified, the wide versions of \code{x} and \code{y} are
correlated (\code{stats::cor(wide_x, wide_y)}) resulting in a correlation matrix with the columns of \code{x} in the
rows and the columns of \code{y} in the columns.

If \code{y} is NULL but \code{rows2}, \code{cols2} and \code{values2} are specified, the columns are taken from \code{x},
creating two matrices to correlate from \code{x}.
}
\examples{
set.seed(123)
cor_in <- data.frame(row = rep(letters[1:10], each = 5),
                     col = rep(LETTERS[1:5], 10),
                     val = rnorm(50))
# Wide format output (default)
corr_wide <- cor_long(cor_in, row, col, val)

# Long format output
corr_long <- cor_long(cor_in, row, col, val,
                      out_format = "long")

# Correlation between two matrices
cor_in2 <- data.frame(rows = rep(letters[1:10], each = 10),
                      cols = rep(letters[1:10], 10),
                      values = rnorm(100))
corr2 <- cor_long(cor_in, row, col, val,
                  cor_in2, rows, cols, values)

}
