% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_shape_long.R
\name{cor_shape_long}
\alias{cor_shape_long}
\title{Convert a correlation matrix (symmetric) to long format}
\usage{
cor_shape_long(x, unique_pairs = F)
}
\arguments{
\item{x}{Correlation matrix to convert (output of \code{stats::cor}).}

\item{unique_pairs}{Whether only unique combinations should be included in the output.}
}
\value{
A data frame with the columns 'row', 'column' (indicating combinations), and 'cor' (correlation coefficients).
}
\description{
Convert a correlation matrix (symmetric) to long format
}
\examples{
cor_shape_long(cor(mtcars))
}
