% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{default_col_scale}
\alias{default_col_scale}
\title{Get default colour scale for non-correlation heatmaps.}
\usage{
default_col_scale(
  val_type,
  aes_type,
  leg_order = 1,
  title = ggplot2::waiver(),
  na_col = "grey50",
  bins = NULL,
  limits = NULL
)
}
\arguments{
\item{val_type}{Value type ('continuous' or 'discrete').}

\item{aes_type}{Aesthetic type ('fill', 'col' or 'size).}

\item{leg_order}{Order of legend.}

\item{title}{Legend title.}

\item{na_col}{Colour of NA values.}

\item{bins}{Number of bins in scale(s).}

\item{limits}{Limits of scales.}
}
\value{
ggplot2 colour scale for non-correlation heatmaps.
}
\description{
Get default colour scale for non-correlation heatmaps.
}
\keyword{internal}
