% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{default_scale_corr}
\alias{default_scale_corr}
\title{Get a default scale for correlation heatmaps.}
\usage{
default_scale_corr(
  aes_type,
  bins = NULL,
  limits = c(-1, 1),
  size_range = NULL,
  high = "sienna2",
  mid = "white",
  low = "skyblue2",
  midpoint = 0,
  na_col = "grey50",
  leg_order = 1,
  title = ggplot2::waiver()
)
}
\arguments{
\item{aes_type}{Type of aesthetic ('fill', 'col', or 'size').}

\item{bins}{Number of bins (for fill and colour scales).}

\item{limits}{Scale limits (fill and colour).}

\item{size_range}{Range of sizes (size).}

\item{high}{Colours at higher end of fill or colour scale.}

\item{mid}{Colours at middle point of fill or colour scale.}

\item{low}{Colours at lower end of fill or colour scale.}

\item{midpoint}{Middle point of colour scale.}

\item{na_col}{Colour of NAs.}

\item{leg_order}{Order of legend.}

\item{title}{Title of legend.}
}
\value{
ggplot2 scale for correlation heatmap.
}
\description{
Get a default scale for correlation heatmaps.
}
\keyword{internal}
