% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{default_size_scale}
\alias{default_size_scale}
\title{Get default size scale for non-correlation heatmaps.}
\usage{
default_size_scale(val_type, leg_order = 1, title = ggplot2::waiver())
}
\arguments{
\item{val_type}{Value type ('continuous' or 'discrete').}

\item{leg_order}{Order of legend.}

\item{title}{Legend title.}
}
\value{
ggplot2 size scale for non-correlation heatmaps.
}
\description{
Get default size scale for non-correlation heatmaps.
}
\keyword{internal}
