% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{default_size_scale_corr}
\alias{default_size_scale_corr}
\title{Get a default size scale for correlation heatmaps.}
\usage{
default_size_scale_corr(
  size_range = NULL,
  leg_order = 1,
  title = ggplot2::waiver()
)
}
\arguments{
\item{size_range}{Numeric vector of length 1 or 2 for size limits.}

\item{leg_order}{Order of legend.}

\item{title}{Title of legend.}
}
\value{
ggplot2 size scale for correlation heatmap (absolute value transform).
}
\description{
Get a default size scale for correlation heatmaps.
}
\keyword{internal}
