% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation_pos.R
\name{get_annotation_coord}
\alias{get_annotation_coord}
\title{Calculate positions of annotations for heatmap}
\usage{
get_annotation_coord(
  annot_dim = c("rows", "cols"),
  annot_side = T,
  annot_names,
  data_size
)
}
\arguments{
\item{data_size}{}
}
\value{
Numeric vector of annotation cell positions
}
\description{
Calculate positions of annotations for heatmap
}
