% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{get_annotation_pos}
\alias{get_annotation_pos}
\title{Calculate positions of annotations for heatmap.}
\usage{
get_annotation_pos(
  annot_side = TRUE,
  annot_names,
  annot_size,
  annot_dist,
  annot_gap,
  data_size
)
}
\arguments{
\item{annot_side}{Logical specifying annotation position. TRUE is left of the heatmap if row annotation,
bottom of heatmap if column annotation}

\item{annot_names}{Names of the annotations.}

\item{annot_size}{Size of annotation cells where 1 is the size of a heatmap cell.}

\item{annot_dist}{Distance between heatmap and first annotation.}

\item{annot_gap}{Size of gap between annotations.}

\item{data_size}{Number of rows or columns in the main heatmap (which is symmetric). Used to get
starting position of annotations.}
}
\value{
Numeric vector of annotation cell positions
}
\description{
Calculate positions of annotations for heatmap.
}
\keyword{internal}
