% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{get_colour_scale}
\alias{get_colour_scale}
\title{Colour scale dispenser.}
\usage{
get_colour_scale(num, type)
}
\arguments{
\item{num}{Integer between 1 and 8.}

\item{type}{String specifying class.}
}
\value{
String with brewer or viridis scale.
}
\description{
Colour scale dispenser.
}
\keyword{internal}
