% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{get_default_annot_scale}
\alias{get_default_annot_scale}
\title{Colour scale dispenser.}
\usage{
get_default_annot_scale(num, type)
}
\arguments{
\item{num}{Integer between 1 and 8.}

\item{type}{String specifying data type (discrete or continuous).}
}
\value{
Brewer or Viridis ggplot2 scale name.
}
\description{
Colour scale dispenser.
}
\keyword{internal}
