% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_corr_heatmap.R
\name{gg_corr_heatmap}
\alias{gg_corr_heatmap}
\title{Make a correlation heatmap with ggplot2.}
\usage{
gg_corr_heatmap(
  data,
  cor_method = "pearson",
  cor_use = "everything",
  high = "sienna2",
  mid = "white",
  low = "skyblue2",
  limits = c(-1, 1),
  layout = "lowerright",
  include_diag = F,
  return_data = F,
  cell_shape = "heatmap",
  label_cells = F,
  cell_label_size = 3,
  cell_label_digits = 2,
  border_col = "grey",
  border_lwd = 0.5,
  names_diag = T,
  params_diag = NULL,
  names_x = F,
  names_x_side = "top",
  names_y = F,
  names_y_side = "left",
  annot_rows_df = NULL,
  annot_cols_df = NULL,
  annot_rows_legend = T,
  annot_cols_legend = T,
  annot_rows_colr = annot_colr,
  annot_cols_colr = annot_colr,
  annot_rows_side = "right",
  annot_cols_side = "bottom",
  annot_dist = 0.2,
  annot_rows_dist = annot_dist,
  annot_cols_dist = annot_dist,
  annot_gap = 0,
  annot_rows_gap = annot_gap,
  annot_cols_gap = annot_gap,
  annot_size = 0.5,
  annot_rows_size = annot_size,
  annot_cols_size = annot_size,
  annot_label = T,
  annot_rows_label = annot_label,
  annot_cols_label = annot_label,
  annot_label_size = 4,
  annot_rows_label_size = annot_label_size,
  annot_cols_label_size = annot_label_size,
  annot_rows_border_col = border_col,
  annot_cols_border_col = annot_rows_border_col,
  annot_rows_border_lwd = border_lwd,
  annot_cols_border_lwd = annot_rows_border_lwd,
  clust_data = F,
  dist_method = "euclidean",
  clust_method = "complete",
  dend_rows = T,
  dend_cols = T,
  dend_rows_height = 0.3,
  dend_cols_height = 0.3,
  dend_rows_lwd = 0.3,
  dend_cols_lwd = 0.3,
  dend_rows_side = "right",
  dend_cols_side = "bottom",
  dend_rows_colour = "black",
  dend_cols_colour = "black",
  dend_options = NULL,
  legend_position = "right",
  plot_margin = c(20, 10, 10, 20),
  margin_unit = "pt"
)
}
\arguments{
\item{data}{A matrix or data frame containing numeric values, where the columns will be correlated with each other.}

\item{cor_method}{String specifying correlation method to use in the \code{cor} function. Default is 'pearson'.}

\item{cor_use}{String specifying the \code{use} argument of \code{cor}, which defineshow to deal with missing values. Default is 'everything'.}

\item{high}{Name of the colour to use for the highest value of the colour scale.}

\item{mid}{Name of the colour to use for 0 in the colour scale.}

\item{low}{Name of the colour to use for the lowest value of the colour scale.}

\item{limits}{Correlation limits to plot between.}

\item{layout}{String specifying the layout of the output correlation heatmap. Possible layouts include
top left, top right, bottom left, bottom right (default), or the whole heatmap. The string should be composed
of the vertical position (top or bottom) followed by the horizontal position (left or right). Bottom can
be specified by 'bottom', 'lower', 'down', or the first letter of these. Left is specified by 'left' or 'l'.
'full', 'whole', or 'all' (or 'f', 'w', 'a') result in the whole correlation matrix being plotted.
For any other strings top and right are selected.}

\item{include_diag}{Logical indicating if the diagonal cells should be plotted (included either way if the whole matrix is plotted).}

\item{return_data}{Logical indicating if the data used for plotting (i.e. the correlation values) should be returned.}

\item{cell_shape}{Value specifying what shape the heatmap cells should take. Any non-numeric value will result in a normal heatmap with square cells (default).
A numeric value can be used to specify an R shape (pch) to use, such as 21 for filled circles. Note that only shapes 21-25 support filling (others will not display the heatmap colour properly).}

\item{label_cells}{Logical specifying if the cells should be labelled with the correlation values.}

\item{cell_label_size}{Size of cell labels, used as the \code{size} argument in \code{ggplot2::geom_text}.}

\item{cell_label_digits}{Number of digits to display when cells are labelled with correlation coefficients. Default is 2, passed to \code{round}.}

\item{border_col}{Colour of cell borders.}

\item{border_lwd}{Size of cell borders, used for the \code{size} argument in \code{ggplot2::geom_tile}. Set to 0 to remove cell borders.}

\item{names_diag}{Logical indicating if names should be written in the diagonal cells.}

\item{params_diag}{List with named parameters (such as size, angle, etc) passed on to geom_text when writing the column names in the diagonal.}

\item{names_x}{Logical indicating if names should be written on the x axis. Labels can be customised using \code{ggplot2::theme()} on the output plot.}

\item{names_x_side}{String specifying position of the x axis names ("top" or "bottom").}

\item{names_y}{Logical indicating if names should be written on the y axis.}

\item{names_y_side}{String specifying position of the y axis names ("left" or "right").}

\item{annot_rows_df}{Data frame for row annotations. The names of the columns in the data must be included,
either as row names or in a column named \code{.names}. Each other column specifies an annotation where the column name
will be used as the annotation name (in the legend and next to the annotation). Numeric columns will use a continuous
colour scale while factor or character columns use discrete scales.}

\item{annot_cols_df}{Same usage as \code{annot_rows_df} but for column annotation.}

\item{annot_rows_legend}{Logical indicating if a legend should be drawn for row annotations.}

\item{annot_cols_legend}{Logical indicating if a legend should be drawn for column annotations.}

\item{annot_rows_colr}{Named list for row annotation colour scales. The names should specify which annotation the scales apply to.
Elements can be strings or ggplot2 "Scale" class objects. If a string it is used as the brewer palette (categorical annotation) or viridis option (continuous annotation).
If a scale object it is used as is, allowing more flexibility. This may change the order that legends are drawn in,
specify order using the \code{guide} argument in the \code{ggplot2} scale function provided.}

\item{annot_cols_colr}{Named list used for column annotation colour scales, used like \code{annot_rows_colr}.}

\item{annot_rows_side}{String specifying which side row annotation should be drawn ('left' for left, otherwise right).}

\item{annot_cols_side}{String specifying which side column annotation should be drawn ('bottom', 'down', 'lower' for bottom, otherwise top).}

\item{annot_dist}{Distance between heatmap and first annotation cell where 1 is the size of one heatmap cell. Used for both row and column annotation.}

\item{annot_rows_dist}{Used to overwrite \code{annot_dist} for row annotation.}

\item{annot_cols_dist}{Used to overwrite \code{annot_dist} for column annotation.}

\item{annot_gap}{Distance between each annotation where 1 is the size of one heatmap cell. Used for both row and column annotation.}

\item{annot_rows_gap}{Used to overwrite \code{annot_gap} for row annotation.}

\item{annot_cols_gap}{Used to overwrite \code{annot_gap} for column annotation.}

\item{annot_size}{Size (width for row annotation, height for column annotation) of annotation cells. Used for both row and column annotation.}

\item{annot_rows_size}{Used to overwrite \code{annot_size} for row annotation.}

\item{annot_cols_size}{Used to overwrite \code{annot_size} for column annotation.}

\item{annot_label}{not yet implemented (control if names of annotations should be shown in drawing area)}

\item{annot_rows_label}{not yet implemented}

\item{annot_cols_label}{not yet implemented}

\item{annot_label_size}{not yet implemented}

\item{annot_rows_label_size}{not yet implemented}

\item{annot_cols_label_size}{not yet implemented}

\item{annot_rows_border_col}{Colour of row annotation borders. By default uses the same colour as the heatmap cell borders.}

\item{annot_cols_border_col}{Colour of column annotation borders. By default uses the same colour as the row annotation cell borders.}

\item{annot_rows_border_lwd}{Line width of row annotation borders. By default uses the same value as the heatmap cell borders.}

\item{annot_cols_border_lwd}{Line width of column annotation borders. By default uses the same value as the row annotation cell borders.}

\item{clust_data}{Logical indicating if the correlations should be clustered in the heatmap. May also be
a \code{hclust} object where the correlations have been clustered already.}

\item{dist_method}{String with the distance metric to use for clustering, given to \code{dist}.}

\item{clust_method}{String with the clustering method to use, given to \code{hclust}.}

\item{dend_rows}{Logical indicating if a dendrogram should be drawn for the rows.}

\item{dend_cols}{Logical indicating if a dendrogram should be drawn for the columns.}

\item{dend_rows_height}{Number to scale the height of the row dendrogram by when plotting.}

\item{dend_cols_height}{Number to scale the height of the column dendrogram by when plotting.}

\item{dend_rows_lwd}{Line width of the row dendrogram.}

\item{dend_cols_lwd}{Line width of the column dendrogram.}

\item{dend_rows_side}{Which side to draw the row dendrogram on ('left' for left, otherwise right).}

\item{dend_cols_side}{Which side to draw the column dendrogram on ('bottom', 'down', 'lower' for bottom, otherwise top).}

\item{dend_rows_colour}{Colour of the row dendrogram.}

\item{dend_cols_colour}{Colour of the column dendrogram.}

\item{dend_options}{Options to give to \code{dendextend::set} for customising the dendrogram. Only works with options for colour, line type and line width.
The options are specified in a list where each element is a list corresponding to one call of
dendextend::set(). Each such list must either contain the 'what' argument as the first element or
as an element named 'what'. Other elements must be named and sepcify other parameters, such as
'k' for 'branches_k_color' or 'value' used to specify colours.}

\item{legend_position}{Position of the legends (use how???)}

\item{plot_margin}{Plot margins, specified as a numeric vector of length 4 in the order of top, right, bottom, left.}

\item{margin_unit}{Unit to use for the specified margin.}
}
\value{
The correlation heatmap as a \code{ggplot} object.
If \code{return_data} is TRUE the output is a list containing the plot (named 'plot'),
the correlations ('plot_data'), and the result of the clustering ('clustering', only if \code{clust_data} is TRUE).
}
\description{
Make a correlation heatmap with ggplot2.
}
\examples{
gg_corr_heatmap(mtcars)
}
