% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrheatmap-package.R
\docType{package}
\name{ggcorrheatmap-package}
\alias{ggcorrheatmap}
\alias{ggcorrheatmap-package}
\title{ggcorrheatmap: Flexible 'ggplot' Correlation Heatmaps}
\description{
Create flexible 'ggplot2' heatmaps and correlation heatmaps with clustering and annotation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leod123/ggcorrheatmap}
  \item Report bugs at \url{https://github.com/leod123/ggcorrheatmap/issues}
}

}
\author{
\strong{Maintainer}: Leo Dahl \email{leokosdah@gmail.com} [copyright holder]

}
\keyword{internal}
