% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghm.R
\name{gghm}
\alias{gghm}
\title{Make a correlation heatmap with ggplot2.}
\usage{
gghm(
  x,
  fill_scale = NULL,
  fill_name = "value",
  layout = "full",
  include_diag = F,
  return_data = F,
  cell_shape = "heatmap",
  label_cells = F,
  cell_label_size = 3,
  cell_label_digits = NULL,
  border_col = "grey",
  border_lwd = 0.5,
  names_diag = T,
  names_diag_param = NULL,
  names_x = F,
  names_x_side = "top",
  names_y = F,
  names_y_side = "left",
  annot_rows_df = NULL,
  annot_cols_df = NULL,
  annot_rows_fill = NULL,
  annot_cols_fill = NULL,
  annot_rows_side = "right",
  annot_cols_side = "bottom",
  annot_legend = T,
  annot_dist = 0.2,
  annot_gap = 0,
  annot_size = 0.5,
  annot_label = T,
  annot_rows_params = NULL,
  annot_cols_params = NULL,
  annot_rows_label_side = "bottom",
  annot_cols_label_side = "left",
  annot_rows_label_params = NULL,
  annot_cols_label_params = NULL,
  cluster_rows = F,
  cluster_cols = F,
  cluster_distance = "euclidean",
  cluster_method = "complete",
  dend_rows = T,
  dend_cols = T,
  dend_rows_side = "right",
  dend_cols_side = "bottom",
  dend_col = "black",
  dend_height = 0.3,
  dend_lwd = 0.3,
  dend_lty = 1,
  dend_rows_params = NULL,
  dend_cols_params = NULL,
  dend_rows_extend = NULL,
  dend_cols_extend = NULL,
  legend_position = "right",
  plot_margin = c(20, 10, 10, 20),
  margin_unit = "pt"
)
}
\arguments{
\item{x}{Matrix or data frame in wide format containing the columns to correlate against each other or against the columns in \code{y}.}

\item{layout}{String specifying the layout of the output correlation heatmap. Possible layouts include
top left, top right, bottom left, bottom right (default), or the whole heatmap. The string should be composed
of the vertical position (top or bottom) followed by the horizontal position (left or right). Bottom can
be specified by 'bottom', 'lower', 'down', or the first letter of these. Left is specified by 'left' or 'l'.
'full', 'whole', or 'all' (or 'f', 'w', 'a') result in the whole correlation matrix being plotted.
For any other strings top and right are selected.}

\item{include_diag}{Logical indicating if the diagonal cells should be plotted (included either way if the whole matrix is plotted).}

\item{return_data}{Logical indicating if the data used for plotting (i.e. the correlation values) should be returned.}

\item{cell_shape}{Value specifying what shape the heatmap cells should take. Any non-numeric value will result in a normal heatmap with square cells (default).
A numeric value can be used to specify an R shape (pch) to use, such as 21 for filled circles. Note that only shapes 21-25 support filling (others will not display the heatmap colour properly).}

\item{label_cells}{Logical specifying if the cells should be labelled with the correlation values.}

\item{cell_label_size}{Size of cell labels, used as the \code{size} argument in \code{ggplot2::geom_text}.}

\item{cell_label_digits}{Number of digits to display when cells are labelled with correlation coefficients. Default is 2, passed to \code{round}. NULL for no rounding (or if labels are characters).}

\item{border_col}{Colour of cell borders.}

\item{border_lwd}{Size of cell borders, used for the \code{size} argument in \code{ggplot2::geom_tile}. Set to 0 to remove cell borders.}

\item{names_diag}{Logical indicating if names should be written in the diagonal cells.}

\item{names_diag_param}{List with named parameters (such as size, angle, etc) passed on to geom_text when writing the column names in the diagonal.}

\item{names_x}{Logical indicating if names should be written on the x axis. Labels can be customised using \code{ggplot2::theme()} on the output plot.}

\item{names_x_side}{String specifying position of the x axis names ("top" or "bottom").}

\item{names_y}{Logical indicating if names should be written on the y axis.}

\item{names_y_side}{String specifying position of the y axis names ("left" or "right").}

\item{annot_rows_df}{Data frame for row annotations. The names of the columns in the data must be included,
either as row names or in a column named \code{.names}. Each other column specifies an annotation where the column name
will be used as the annotation name (in the legend and next to the annotation). Numeric columns will use a continuous
colour scale while factor or character columns use discrete scales.}

\item{annot_cols_df}{Same usage as \code{annot_rows_df} but for column annotation.}

\item{annot_rows_fill}{Named list for row annotation colour scales. The names should specify which annotation the scales apply to.
Elements can be strings or ggplot2 "Scale" class objects. If a string it is used as the brewer palette (categorical annotation) or viridis option (continuous annotation).
If a scale object it is used as is, allowing more flexibility. This may change the order that legends are drawn in,
specify order using the \code{guide} argument in the \code{ggplot2} scale function provided.}

\item{annot_cols_fill}{Named list used for column annotation colour scales, used like \code{annot_rows_fill}.}

\item{annot_rows_side}{String specifying which side row annotation should be drawn ('left' for left, otherwise right).}

\item{annot_cols_side}{String specifying which side column annotation should be drawn ('bottom', 'down', 'lower' for bottom, otherwise top).}

\item{annot_legend}{Logical indicating if row and column annotations should have legends.}

\item{annot_dist}{Distance between heatmap and first annotation cell where 1 is the size of one heatmap cell. Used for both row and column annotation.}

\item{annot_gap}{Distance between each annotation where 1 is the size of one heatmap cell. Used for both row and column annotation.}

\item{annot_size}{Size (width for row annotation, height for column annotation) of annotation cells. Used for both row and column annotation.}

\item{annot_label}{not yet implemented (control if names of annotations should be shown in drawing area)}

\item{annot_rows_params}{Named list with parameters for row annotations to overwrite the defaults set by the \verb{annot_*} arguments, each name corresponding to the \code{*} part
(see details for more information).}

\item{annot_cols_params}{Named list with parameters for column annotations, used like \code{annot_rows_params}.}

\item{annot_rows_label_side}{String specifying which side the row annotation labels should be on. Either "top" or "bottom".}

\item{annot_cols_label_side}{String specifying which side the column annotation labels should be on. Either "left" or "right".}

\item{annot_rows_label_params}{Named list of parameters for row annotation labels. Given to \code{grid::textGrob}, see \code{?grid::textGrob} for details. \code{?grid::gpar} is also helpful.}

\item{annot_cols_label_params}{Named list of parameters for column annotation labels. Given to \code{grid::textGrob}, see \code{?grid::textGrob} for details. \code{?grid::gpar} is also helpful.}

\item{cluster_rows}{Logical indicating if rows should be clustered.}

\item{cluster_cols}{Logical indicating if columns should be clustered.}

\item{cluster_distance}{String with the distance metric to use for clustering, given to \code{dist}.}

\item{cluster_method}{String with the clustering method to use, given to \code{hclust}.}

\item{dend_rows}{Logical indicating if a dendrogram should be drawn for the rows.}

\item{dend_cols}{Logical indicating if a dendrogram should be drawn for the columns.}

\item{dend_rows_side}{Which side to draw the row dendrogram on ('left' for left, otherwise right).}

\item{dend_cols_side}{Which side to draw the column dendrogram on ('bottom', 'down', 'lower' for bottom, otherwise top).}

\item{dend_col}{Colour to use for dendrogram lines, applied to both row and column dendrograms.}

\item{dend_height}{Number by which to scale dendrogram height, applied to both row and column dendrograms.}

\item{dend_lwd}{Linewidth of dendrogram lines, applied to both row and column dendrograms.}

\item{dend_lty}{Dendrogram line type, applied to both row and column dendrograms.}

\item{dend_rows_params}{Named list for row dendrogram parameters. See details for more information.}

\item{dend_cols_params}{Named list for column dendrogram parameters. See details for more information.}

\item{dend_rows_extend}{Named list for specifying \code{dendextend} functions to apply to the row dendrogram. See details for usage.}

\item{dend_cols_extend}{Named list for specifying \code{dendextend} functions to apply to the column dendrogram. See details for usage.}

\item{legend_position}{Position of the legends, given to \code{ggplot2::theme}. Either a string for the position outside the plotting area,
or a numeric vector of length 2 for normalised coordinates inside the plotting area.}

\item{plot_margin}{Plot margins, specified as a numeric vector of length 4 in the order of top, right, bottom, left.}

\item{margin_unit}{Unit to use for the specified margin.}

\item{y}{Optional matrix or data frame in wide format containing columns to correlate with the columns in \code{x}.}

\item{cor_method}{String specifying correlation method to use in the \code{cor} function. Default is 'pearson'.}

\item{cor_use}{String specifying the \code{use} argument of \code{cor}, which defineshow to deal with missing values. Default is 'everything'.}

\item{high}{Name of the colour to use for the highest value of the colour scale.}

\item{mid}{Name of the colour to use for 0 in the colour scale.}

\item{low}{Name of the colour to use for the lowest value of the colour scale.}

\item{limits}{Correlation limits to plot between.}

\item{bins}{Specify number of bins if the correlation scale should be binned. NULL for a continuous scale.}
}
\value{
The correlation heatmap as a \code{ggplot} object.
If \code{return_data} is TRUE the output is a list containing the plot (named 'plot'),
the correlations ('plot_data'), and the result of the clustering ('clustering', only if \code{cluster_data} is TRUE).
}
\description{
Make a correlation heatmap with ggplot2.
}
\details{
The annotation parameter arguments \code{annot_rows_params} and \code{annot_cols_params} should be named lists, where the possible options correspond to
the different \verb{annot_*} arguments. The possible options are "legend" (logical, if legends should be drawn), "dist" (distance between heatmap and annotation), "gap" (distance between annotations),
"size" (cell size), "label" (logical, if the annotation names should be displayed), and "label_size" (label text size). Any unused options will
use the defaults set by the \verb{annot_*} arguments.

The dendrogram parameters arguments \code{dend_rows_params} and \code{dend_cols_params} should be named lists, analogous to the annotation parameter arguments. Possible options are
"col" (line colour), "height" (height scaling), "lwd" (line width), and "lty" (line type).

The \code{dend_options} argument makes it possible to customise the dendrograms using the \code{dendextend} package.
The argument should be a named list, each element named after the \code{dendextend} function to use (consecutive usage of the \code{set} function
is supported due to duplicate list names being possible). Each element should contain any arguments given to the \code{dendextend} function,
such as the \code{what} argument used in the \code{set} function. See examples for example usage.
}
\examples{
# Basic usage
gg_corr_heatmap(mtcars)

# Different layout
gg_corr_heatmap(mtcars, layout = "f")

# With clustering
gg_corr_heatmap(mtcars, layout = "tl", cluster_data = T)

# With annotation
set.seed(123)
annot <- data.frame(.names = colnames(mtcars),
                    annot1 = rnorm(ncol(mtcars)),
                    annot2 = sample(letters[1:3], ncol(mtcars), T))
gg_corr_heatmap(mtcars, layout = "tr", annot_rows_df = annot,
                # Change margins to fit annotation labels
                plot_margin = c(20, 10, 60, 20))

# Using the dend_options argument
gg_corr_heatmap(mtcars, cluster_data = T, dend_options =
  list("set" = list("branches_lty", c(1, 2, 3)),
       "set" = list("branches_k_color", k = 3),
       # Empty list element (or NULL) if no arguments to be given
       "highlight_branches_lwd" = list()))
}
