% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghm_long.R
\name{gghm_tidy}
\alias{gghm_tidy}
\title{gghm for long format data using tidy input.}
\usage{
gghm_tidy(
  x,
  rows,
  cols,
  values,
  labels = NULL,
  annot_rows = NULL,
  annot_cols = NULL,
  ...
)
}
\arguments{
\item{x}{Data frame containing data to plot.}

\item{rows}{Name of column to use as rows.}

\item{cols}{Name of column to use as columns.}

\item{values}{Name of column to use for cell values.}

\item{annot_rows}{Names of columns to use for row annotations.}

\item{annot_cols}{Names of columns to use for column annotations.}

\item{...}{Additional arguments for \code{gghm()}.}
}
\value{
A ggplot2 object with the heatmap. If \code{return_data} is \code{TRUE}, plotting data is returned as well.
}
\description{
gghm for long format data using tidy input.
}
