% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{increment1to8}
\alias{increment1to8}
\title{Increment between 1 and 8.}
\usage{
increment1to8(x)
}
\arguments{
\item{x}{Integer to increment.}
}
\value{
Integer. x + 1 if x is between 1 and 7, 1 otherwise.
}
\description{
Increment between 1 and 8.
}
\keyword{internal}
