% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{isSquare}
\alias{isSquare}
\title{Check if a matrix is square (and require the same row and colnames).}
\usage{
isSquare(x)
}
\arguments{
\item{x}{A matrix or data frame to check.}
}
\value{
Logical.
}
\description{
A 'softer' version of isSymmetric that does not require the values to be the same,
only the dimensions and dimnames.
}
\keyword{internal}
