% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_facets.R
\name{make_facet_vector}
\alias{make_facet_vector}
\title{Make vector for facetting}
\usage{
make_facet_vector(facet_in, len)
}
\arguments{
\item{facet_in}{User input for facetting (split_rows, split_cols).}

\item{len}{Length of output vector.}
}
\value{
Vector of facet memberships.
}
\description{
Make vector for facetting
}
\keyword{internal}
