% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{make_legend_order}
\alias{make_legend_order}
\title{Make legend order depending on what the plot will contain.}
\usage{
make_legend_order(
  mode,
  col_scale = NULL,
  size_scale = NULL,
  annot_rows_df = NULL,
  annot_cols_df = NULL,
  legend_order = NULL
)
}
\arguments{
\item{mode}{Plotting modes.}

\item{col_scale}{One or two colour scales (shared for fill and colour). NULL for default, string for Brewer or Viridis, or a scale.}

\item{size_scale}{Size scales (NULL or ggplot2 scales).}

\item{annot_rows_df}{Annotation data frame for rows.}

\item{annot_cols_df}{Annotation data frame for columns.}

\item{legend_order}{Numeric vector with legend order. NULL for default.}
}
\value{
A list with aesthetics for the main plot and orders of all legends.
}
\description{
Make legend order depending on what the plot will contain.
}
\keyword{internal}
