% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{mirror_dendrogram}
\alias{mirror_dendrogram}
\title{Mirror coordinates of dendrogram around its middle point}
\usage{
mirror_dendrogram(dend_seg, dend_dim = c("rows", "cols"))
}
\arguments{
\item{dend_seg}{Data frame containing dendrogram segments, attained from \code{dendextend::as.ggdend()}}

\item{dend_dim}{Character specifying whether the dendrogram is linked to rows or columns in the heatmap}
}
\value{
Dendrogram segment coordinates dataframe, with the specified dimension mirrored
}
\description{
Mirror a dendrogram around an axis parallel to the height, located in its middle point
}
\keyword{internal}
