% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_dendrogram.R
\name{move_dendrogram}
\alias{move_dendrogram}
\title{Move coordinates of dendrogram to edges of heatmap}
\usage{
move_dendrogram(
  cor_long,
  dend_seg,
  dend_dim = c("rows", "cols"),
  dend_pos,
  annot_df,
  annot,
  annot_pos,
  annot_coord,
  annot_size
)
}
\arguments{
\item{cor_long}{Long format data frame with the correlations}

\item{dend_seg}{Data frame containing dendrogram segments, attained from \code{dendextend::as.ggdend()}}

\item{dend_dim}{Character specifying whether the dendrogram is linked to rows or columns in the heatmap}

\item{dend_pos}{Logical specifying dendrogram position. TRUE is left of the heatmap if row dendrogram, top of heatmap if column dendrogram}

\item{annot_df}{Data frame with annotations for checking that annotations exist as well as their size}

\item{annot}{Logical indicating if annotations are intended to be drawn or not}

\item{annot_pos}{Logical specifying annotation position, analogous to \code{dend_pos}}

\item{annot_coord}{Numeric vector of annotation coordinates (x coordinates for row annotations, y for column annotations)}

\item{annot_size}{Numeric of length 1, the specified size (width or height) of annotation cells}
}
\value{
Data frame with updated dendrogram coordinates
}
\description{
Move coordinates of dendrogram by calculating distance it has to move to end up at the desired edges of the heatmap
}
\keyword{internal}
