% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{orient_dendrogram}
\alias{orient_dendrogram}
\title{Orient a dendrogram.}
\usage{
orient_dendrogram(
  dend,
  dim = c("rows", "cols"),
  full_plt,
  layout,
  dend_left,
  dend_down
)
}
\arguments{
\item{dend}{Dendrogram segments or nodes data frame (containing x, y, xend, yend).}

\item{dim}{String, rows or cols to know which dimensions dendrogram should be on.}

\item{full_plt}{Logical indicating if it's for the full layout.}

\item{dend_left}{Logical indicating if the dendrogram should be placed on the left (only relevant if row dend).}

\item{dend_down}{Logical indicating if the dendrogram should be placed at the bottom (only relevant if col dend).}
}
\value{
The input dendrogram data frame but rotated and mirrored to fit the plot.
}
\description{
Orient a dendrogram.
}
\keyword{internal}
