% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{prepare_annot_col}
\alias{prepare_annot_col}
\title{Prepare default colour scales for annotation.}
\usage{
prepare_annot_col(
  annot_rows_df = NULL,
  annot_cols_df = NULL,
  annot_rows_col = NULL,
  annot_cols_col = NULL
)
}
\arguments{
\item{annot_rows_df}{Data frame with annotation for rows.}

\item{annot_cols_df}{Data frame with annotation for columns.}

\item{annot_rows_col}{List with colour scales for rows.}

\item{annot_cols_col}{List with colour scales for columns.}
}
\value{
List of length two containing lists of row annotation and column annotation.
}
\description{
Prepares a brewer palette or viridis option for all annotations that don't have any colour scale
specified by the user. There are eight options each for brewer (categorical) and viridis (continuous) and they are selected
sequentially, going back to the beginning if there are more than eight annotations of each kind.
}
\keyword{internal}
