% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{prepare_annotation}
\alias{prepare_annotation}
\title{Prepare annotation parameters and positions}
\usage{
prepare_annotation(
  annot_df,
  annot_defaults,
  annot_params,
  lannot_side,
  annot_label_params,
  annot_label_side,
  data_size
)
}
\arguments{
\item{annot_df}{Annotation data frame. Should contain the rownames or a column called '.names' with names, other columns are used for annotation}

\item{annot_defaults}{Default parameters for annotation.}

\item{annot_params}{Provided annotation parameters to update.}

\item{lannot_side}{Logical indicating annotation side. TRUE if left (row annotations) or bottom (column annotations).}

\item{annot_label_params}{Annotation label (names next to annotations) parameters to update.}

\item{annot_label_side}{Annotation label side.}

\item{data_size}{Size of data (ncol if row annotations, nrow if column annotations).}
}
\value{
List with updated annotation parameters, calculated annotation positions, and updated
annotation label parameters.
}
\description{
Prepare annotation parameters and positions
}
\keyword{internal}
