% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrhm.R
\name{prepare_cell_labels}
\alias{prepare_cell_labels}
\title{Prepare cell labels in ggcorrhm to pass to gghm}
\usage{
prepare_cell_labels(
  mode,
  cell_labels,
  p_values,
  cell_label_p,
  cell_label_digits,
  p_thresholds = NULL,
  cor_mat_dat = NULL
)
}
\arguments{
\item{mode}{Plotting mode.}

\item{cell_labels}{Cell labels (TRUE, FALSE, matrix, data frame).}

\item{p_values}{Logical indicating if p-values should be computed.}

\item{cell_label_p}{Logical indicating if cell labels should be swapped for p-values.}

\item{cell_label_digits}{Number of digits to display for cell labels.}

\item{p_thresholds}{P-value thresholds.}

\item{cor_mat_dat}{Correlation long format data from correlation tests.}
}
\value{
Object to use as cell_labels in gghm (containing a logical, a matrix/data frame, or a list of length 2 with those things).
}
\description{
Prepare cell labels in ggcorrhm to pass to gghm
}
\keyword{internal}
