% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_facets.R
\name{prepare_facets}
\alias{prepare_facets}
\title{Prepare facetting columns in the plotting data.}
\usage{
prepare_facets(
  x_long,
  x,
  facet_in,
  layout,
  context = c("row", "col"),
  dendro = NULL
)
}
\arguments{
\item{x_long}{Long format plotting data.}

\item{x}{Wide format plotting data.}

\item{facet_in}{Facetting user input (vector or data frame).}

\item{layout}{Plot layout vector.}

\item{context}{Context of facetting (row or col).}

\item{dendro}{List containing clustering results (or NULL if no clustering).}
}
\value{
Long format plotting data with facetting column added.
}
\description{
Prepare facetting columns in the plotting data.
}
\keyword{internal}
