% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_facets.R
\name{prepare_facets_tidy}
\alias{prepare_facets_tidy}
\title{Prepare facets for gghm_tidy}
\usage{
prepare_facets_tidy(x, id_col, facet_col, params, context = c("row", "column"))
}
\arguments{
\item{x}{Input long format data frame.}

\item{id_col}{Column containing IDs (rows or column names).}

\item{facet_col}{Column containing facet memberships.}

\item{params}{Input arguments for gghm (passed to ...).}

\item{context}{Context for facets (row or column).}
}
\value{
Vector containing facet membersips.
}
\description{
Prepare facets for gghm_tidy
}
\keyword{internal}
