% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrhm.R
\name{prepare_scales}
\alias{prepare_scales}
\title{Prepare scales for mixed scales in ggcorrhm}
\usage{
prepare_scales(
  scale_in,
  type,
  bins = NULL,
  limits = c(-1, 1),
  size_range = NULL,
  high = NULL,
  mid = NULL,
  low = NULL,
  na_col = "grey",
  show_leg
)
}
\arguments{
\item{scale_in}{Scale object from user.}

\item{type}{Type of scale (fill, col, size).}

\item{bins}{Number of bins if binned scale.}

\item{limits}{Scale limits.}

\item{size_range}{Size range}

\item{high}{Colour at higher end of scale.}

\item{mid}{Colour at midpoint.}

\item{low}{Colour at lowest point of scale.}

\item{na_col}{Colour of NAs.}

\item{show_leg}{If the legend should be shown.}
}
\value{
Scale object or list of scale objects.
}
\description{
Prepare scales for mixed scales in ggcorrhm
}
\keyword{internal}
