% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{process_dendrogram}
\alias{process_dendrogram}
\title{Process dendrogram with customisation options}
\usage{
process_dendrogram(mat, dendro, dend_options)
}
\arguments{
\item{mat}{Data that was used for clustering.}

\item{dendro}{Dendrogram object.}

\item{dend_options}{Dendrogram extension options (list or fseq or NULL).}
}
\value{
Processed dendrogram object.
}
\description{
Process dendrogram with customisation options
}
\keyword{internal}
