% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{remove_duplicate_scales}
\alias{remove_duplicate_scales}
\title{Remove duplicate scales.}
\usage{
remove_duplicate_scales(scale_vec, col_scale = NULL, size_scale = NULL)
}
\arguments{
\item{scale_vec}{Vector of scale aesthetics.}

\item{col_scale}{Input colour scales (NULL, string or scale object).}

\item{size_scale}{Input size scales.}
}
\value{
Vector of aesthetics with duplicates removed if appropriate.
}
\description{
If a mixed layout uses the same aesthetic for both triangles and only one (or no) colour or size scale has been specified, remove redundant scales.
}
\keyword{internal}
