% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_default.R
\name{replace_default}
\alias{replace_default}
\title{Replace default elements in a named list with corresponding elements in a new list.}
\usage{
replace_default(default_param, new_param, add_new = F)
}
\arguments{
\item{default_param}{Named list with elements to potentially replace ("defaults").}

\item{new_param}{Named list with elements to replace with.}
}
\value{
A named list where overlapping elements with overlapping names are replaced.
}
\description{
Replace default elements in a named list with corresponding elements in a new list.
}
\examples{
\dontrun{
replace_default(list(a = 1, b = 2, c = 3), list(a = 123, 456, d = 789))
}
}
\keyword{internal}
