% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{rotate_coord}
\alias{rotate_coord}
\title{Rotate cartesian coordinates around 0}
\usage{
rotate_coord(x, angle, radians = T)
}
\arguments{
\item{x}{Matrix of coordinates to rotate. Should be either a numeric vector of length 2, or a matrix with two rows (for rotating multiple coordinates at once)}

\item{angle}{The angle to rotate by. Default is radians.}

\item{radians}{Logical indicating if the given angle is in radians. If FALSE, degrees are used.}
}
\value{
A numeric matrix of rotated coordinates. Contains two rows (x and y coordinates) and one column per input coordinate.
}
\description{
Rotate cartesian coordinates around 0
}
\examples{
rotate_coord(c(1, 2), pi)

rotate_coord(matrix(c(1, 2,
                      3, 4),
                      byrow = FALSE, nrow = 2),
            angle = 90, radians = FALSE)
}
\keyword{internal}
