% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{scale_dendrogram}
\alias{scale_dendrogram}
\title{Scale height of dendrogram.}
\usage{
scale_dendrogram(dend_seg, context = c("rows", "cols"), dend_side, dend_height)
}
\arguments{
\item{dend_seg}{Data frame containing dendrogram segments in format obtained from \code{dendextend}.}

\item{context}{Dimension to draw dendrogram along (rows or columns).}

\item{dend_side}{Logical indicating which side to draw dendrogram on.
If row dendrogram TRUE is left. If column dendrogram TRUE is down.}

\item{dend_height}{Scaling parameter for dendrogram height (1 is no scaling).}
}
\value{
Data frame containing coordinates for dendrogram segments (and any colour, linewidth, line type parameters)
}
\description{
Scale height of dendrogram.
}
\keyword{internal}
