% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghm.R
\name{scale_mat}
\alias{scale_mat}
\title{Scale data rows or columns.}
\usage{
scale_mat(x, scl = NULL)
}
\arguments{
\item{x}{Matrix to scale.}

\item{scl}{String of dimension to scale ("row", "column", "none" (or NULL), or a substring of the beginning).}
}
\value{
Scaled (or not scaled) matrix.
}
\description{
Scale data rows or columns.
}
\keyword{internal}
