% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_mat_long.R
\name{shape_mat_long}
\alias{shape_mat_long}
\title{Convert a matrix to long format using row names and column names.}
\usage{
shape_mat_long(x, unique_pairs = FALSE, na_remove = FALSE)
}
\arguments{
\item{x}{Matrix to convert to long format.}

\item{unique_pairs}{Whether only unique combinations should be included in the output (for symmetric matrices).}

\item{na_remove}{Logical indicating if NAs should be excluded (removes NaNs too).}
}
\value{
A data frame with the columns 'row', 'column' (indicating combinations), and 'value'
}
\description{
Convert a matrix to long format using row names and column names.
}
\keyword{internal}
