% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_mat_long.R
\name{shape_mat_wide}
\alias{shape_mat_wide}
\title{Convert long format matrix to wide.}
\usage{
shape_mat_wide(x)
}
\arguments{
\item{x}{Long format matrix as obtained from 'shape_mat_long' (contains columns named 'row', 'col', and 'value').}
}
\value{
Wide format version of x.
}
\description{
Convert long format matrix to wide.
}
\keyword{internal}
