% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cor.R
\name{test_cor}
\alias{test_cor}
\title{Calculate correlations and p-values between columns.}
\usage{
test_cor(
  x,
  y = NULL,
  method = "pearson",
  use = "everything",
  p_adj_method = "none"
)
}
\arguments{
\item{x}{Matrix or data frame with columns to correlate.}

\item{y}{Matrix or data frame, if provided will be correlated with x.}

\item{method}{Passed to \code{stats::cor}.}

\item{use}{Passed to \code{stats::cor}.}

\item{p_adj_method}{P-value adjustment method, passed to \code{stats::p.adjust}.}
}
\value{
Data frame in long format with correlation values and nominal and adjusted p-values.
}
\description{
Calculate correlations and p-values between columns.
}
\keyword{internal}
