% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{add_diag_names}
\alias{add_diag_names}
\title{Add diagonal names to heatmap.}
\usage{
add_diag_names(plt, x_long, names_diag_params = NULL)
}
\arguments{
\item{plt}{Plot object to add names to.}

\item{x_long}{Long format plotting data.}

\item{names_diag_params}{Parameters for diagonal names.}
}
\value{
Plot with labels added.
}
\description{
Add diagonal names to heatmap.
}
\keyword{internal}
