% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{apply_dendextend}
\alias{apply_dendextend}
\title{Successively apply dendextend functions to a dendrogram.}
\usage{
apply_dendextend(dend_list, dendro)
}
\arguments{
\item{dend_list}{List specifying dendextend functions to apply. For usage see the details of \code{gghm}.}

\item{dendro}{Dendrogram object obtained from \code{stats::as.dendrogram}.}
}
\value{
A dendrogram object modified with dendextend functions.
}
\description{
Successively apply dendextend functions to a dendrogram.
}
\keyword{internal}
