% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendrogram.R
\name{check_dendrogram_pos}
\alias{check_dendrogram_pos}
\title{Check that dendrograms are positioned correctly}
\usage{
check_dendrogram_pos(dat, context = c("row", "col"), dendro)
}
\arguments{
\item{dat}{Long format data for plotting.}

\item{context}{Dimension to which the dendrogram is added. These are used to directly get the columns
in the long data and thus need to be "row" or "col".}

\item{dendro}{The dendrogram segments and nodes list.}
}
\value{
\code{dendro} is returned as is if the positions are correct. Otherwise there is an error.
}
\description{
Check that dendrograms are positioned correctly
}
\keyword{internal}
