% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{get_colour_scale}
\alias{get_colour_scale}
\title{Get a Brewer or Viridis colour scale.}
\usage{
get_colour_scale(
  name,
  val_type,
  aes_type,
  limits = NULL,
  bins = NULL,
  leg_order = 1,
  title = ggplot2::waiver(),
  na_col = "grey50"
)
}
\arguments{
\item{name}{Scale palette/option name.}

\item{val_type}{Value type ('continuous' or 'discrete').}

\item{aes_type}{Aesthetic type ('fill' or 'col').}

\item{limits}{Limits for scale.}

\item{bins}{Number of bins if binned scale.}

\item{leg_order}{Order of legend.}

\item{title}{Legend title.}
}
\value{
ggplot2 scale using Brewer or Viridis.
}
\description{
Get a Brewer or Viridis colour scale.
}
\keyword{internal}
