% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrheatmap-package.R
\docType{package}
\name{ggcorrheatmap-package}
\alias{ggcorrheatmap}
\alias{ggcorrheatmap-package}
\title{ggcorrheatmap: Advanced ggplot Correlation Heatmaps}
\description{
Create correlation heatmaps in ggplot with clustering and annotation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leod123/ggcorrheatmap}
  \item Report bugs at \url{https://github.com/leod123/ggcorrheatmap/issues}
}

}
\author{
\strong{Maintainer}: Leo Dahl \email{leo.dahl@scilifelab.se} (\href{https://orcid.org/0000-0003-1492-3052}{ORCID})

}
\keyword{internal}
