% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrheatmap-package.R
\docType{package}
\name{ggcorrheatmap-package}
\alias{ggcorrheatmap}
\alias{ggcorrheatmap-package}
\title{ggcorrheatmap: Make Flexible 'ggplot2' Correlation Heatmaps}
\description{
Create correlation heatmaps with 'ggplot2' and customise them with flexible annotation and clustering. Symmetric heatmaps can use triangular or mixed layouts, removing redundant information or displaying complementary information in the two halves. There is also support for general heatmaps not displaying correlations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leod123/ggcorrheatmap}
  \item \url{https://leod123.github.io/ggcorrheatmap/}
  \item Report bugs at \url{https://github.com/leod123/ggcorrheatmap/issues}
}

}
\author{
\strong{Maintainer}: Leo Dahl \email{leokosdah@gmail.com} [copyright holder]

}
\keyword{internal}
