% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_hm.R
\name{layout_hm}
\alias{layout_hm}
\title{Layout heatmap data for plotting}
\usage{
layout_hm(x, layout = "f", na_remove = FALSE)
}
\arguments{
\item{x}{Matrix to plot.}

\item{layout}{Layout (full, triangular (topleft, topright, bottomleft, bottomright)).}

\item{na_remove}{Logical indicating if NAs should be removed.}
}
\value{
Long format data for plotting.
}
\description{
Layout heatmap data for plotting
}
\keyword{internal}
