% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{prepare_annot_col}
\alias{prepare_annot_col}
\title{Prepare default colour scales for annotation.}
\usage{
prepare_annot_col(
  annot_rows_df = NULL,
  annot_cols_df = NULL,
  annot_rows_col = NULL,
  annot_cols_col = NULL
)
}
\arguments{
\item{annot_rows_df}{Data frame with annotation for rows.}

\item{annot_cols_df}{Data frame with annotation for columns.}

\item{annot_rows_col}{List with colour scales for rows.}

\item{annot_cols_col}{List with colour scales for columns.}
}
\value{
List of length two containing lists of row annotation and column annotation.
}
\description{
Prepare default colour scales for annotation.
}
\keyword{internal}
