% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{prepare_annotation}
\alias{prepare_annotation}
\title{Prepare annotation parameters and positions}
\usage{
prepare_annotation(
  annot_df,
  annot_defaults,
  annot_params,
  annot_side,
  context = c("rows", "cols"),
  annot_names_size,
  annot_name_params,
  annot_names_side,
  data_size,
  x_long,
  annot_names_strategy = "geom"
)
}
\arguments{
\item{annot_df}{Annotation data frame. Should contain the rownames or a column called '.names' with names, other columns are used for annotation}

\item{annot_defaults}{Default parameters for annotation.}

\item{annot_params}{Provided annotation parameters to update.}

\item{annot_side}{String for annotation side.}

\item{context}{String stating the context ("rows" or "cols").}

\item{annot_names_size}{Size of annotation names.}

\item{annot_name_params}{Annotation label (names next to annotations) parameters to update.}

\item{annot_names_side}{Annotation label side.}

\item{data_size}{Size of data (ncol if row annotations, nrow if column annotations).}

\item{x_long}{Long format plotting data, to see if there are any facets to take into account.}

\item{annot_names_strategy}{String stating how to draw names (geom, grid). To deal with deprecated argument.}
}
\value{
List with updated annotation parameters, calculated annotation positions, and updated
annotation label parameters.
}
\description{
Prepare annotation parameters and positions
}
\keyword{internal}
