% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scales.R
\name{prepare_scales}
\alias{prepare_scales}
\title{Prepare scales for heatmap.}
\usage{
prepare_scales(
  scale_order,
  context = c("gghm", "ggcorrhm"),
  val_type,
  col_scale = NULL,
  col_name = "value",
  size_scale = NULL,
  size_name = "value",
  bins = NULL,
  limits = c(-1, 1),
  high = "sienna2",
  mid = "white",
  low = "skyblue2",
  midpoint = 0,
  size_range = NULL,
  na_col = "grey50"
)
}
\arguments{
\item{scale_order}{List of necessary scales and their orders, as obtained make_legend_order.}

\item{context}{Scale context (gghm or ggcorrhm) for deciding which defaults to use.}

\item{val_type}{String with type of value ('continuous' or 'discrete').}

\item{col_scale}{Colour scales input.}

\item{col_name}{Colour scale names.}

\item{size_scale}{Size scales input.}

\item{size_name}{Size scale names.}

\item{bins}{Number of bins to divide scale into (only for ggcorrhm).}

\item{limits}{Scale limits (for ggcorrhm).}

\item{high}{Colour at higher end of scale (for ggcorrhm).}

\item{mid}{Colour at middle point of scale (for ggcorrhm).}

\item{low}{Colour at Lower end of scale (for ggcorrhm).}

\item{midpoint}{Middle point of scale (for ggcorrhm).}

\item{size_range}{Size range for size scale (for ggcorrhm).}

\item{na_col}{Colour to use for NAs (for ggcorrhm).}
}
\value{
ggplot2 scales for the plot.
}
\description{
Prepare scales for heatmap.
}
\keyword{internal}
